feat: Add Edit Gallery Page
I added the Gallery Page which is only accessible to admins. I added a remove API that removes uploaded image and deletes from the database, and therefore from the gallery page. I also added a feature that blocks un-executive membners from accessible this page. To avoid non-executive members from viewing the page while loading, I returned 'No Permission!' h2 tag when the user is either not logged in or non-executive.